+++ website/app/docs/DateTime/_meta.js
@@ -1,3 +1,4 @@
 export default {
   DateTime: 'DateTime',
+  Duration: 'Duration',
 };